fix: next/root-params erroring when rerendering after action (#82326)
When implementing the check that disallows using `next/root-params` in
server actions, i forgot that when a page is rerendered after an action,
the render is still wrapped with an `actionAsyncStorage` with
`actionStore.isAction === true`. The fix is to also check that we're in
the `'action'` phase, i.e. we haven't switched to rendering yet.

Closes #82302