[Cache Components] Allow span creation while prerendering (#82350)
When creating spans with OTel random values must be created to follow
the OTel spec and is expected in practical implementations. While it is
a loose concession Next.js will allow spans to be created while
prerendering that have ids generated with random values. and by reading
the current time from the system clock.

This is pragmatic but potentially risky because if you are not careful
and for instance wrap a "use cache" function in a span it is possible
you will accidentally pass the span into the cache function causing it
to miss on every invocation due to the random nature of one of it's
arguments.

To defend against this we do check whether `startActiveSpan` is being
passed a Cache Function. However it is easy to circumvent by wrapping it
in an intermediate function or by using startSpan and context.with
yourself.

In the long run we should consider tainting the Span objects created in
these scopes so they cannot be sent through to a cache. I will consider
a follow up PR to land this if we determine taint is likely to be
stabilized in React.

The pratical implementation of this change is to patch the tracer
provider so that it always provides a patched tracer which exits the
workUnitStorage on span start/creation and re-enters the scope on inner
function exectuion.