loader tree: add special segment name to virtual parallel route segments (#82383)
When constructing the loader tree for parallel route segments, we inject
a `children` segment, which could contain things like slot level
layout/error/loading files. We want a way to distinguish between these
virtual segments that exist just for organizational purposes rather than
relying on `children`. This updates the loader tree construction code to
give these more unique names. In the future, we should see if we can
remove the need for this virtual segment if there are no special files
needed at that segment of the tree.