add unstable_isUnrecognizedActionError (#78933)
`unstable_isUnrecognizedActionError` is a new API that lets user code
check if a server action call failed because the action id wasn't
recognized by the server. This usually happens as a result of version
skew between client and server.

Example usage:

 ```ts
 try {
   await myServerAction();
 } catch (err) {
   if (unstable_isUnrecognizedActionError(err)) {
     // The client is from a different deployment than the server.
     // Reloading the page will fix this mismatch.
     window.alert("Please refresh the page and try again");
     return;
   }
 }
 ```
It can also be used to create a special error boundary, like we do in
the tests added here.

Note that this API is not a complete solution to this problem, as it
doesn't allow handling failures for MPA actions (sent without JS). We
might add a more complete API for this in the future to address that.